backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.12
psycopg2-binary
python-dotenv
python-multipart
